rename proj to easy-stocks, configure dev build
