+++ .env.dev
@@ -1,0 +1,7 @@
+API_URL=https://target-mvd-api.herokuapp.com/api/v1
+RELEASE_STORE_FILE=debug-release-key.keystore
+RELEASE_KEY_ALIAS=androiddebugkey
+RELEASE_STORE_PASSWORD=android
+RELEASE_KEY_PASSWORD=android
+
+
