--- .env.sample
@@ -1,5 +1,0 @@
-API_URL=https://target-mvd-api.herokuapp.com/api/v1
-RELEASE_STORE_FILE=
-RELEASE_KEY_ALIAS=
-RELEASE_STORE_PASSWORD=
-RELEASE_KEY_PASSWORD=
