+++ android/app/_BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.reactnativebase",
+    package = "com.bvapps.easystocks",
 )
 
 android_resource(
     name = "res",
-    package = "com.reactnativebase",
+    package = "com.bvapps.easystocks",
     res = "src/main/res",
 )
