+++ android/app/src/dev/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">ReactNativeBase-Develop</string>
+    <string name="app_name">EasyStocks-Develop</string>
 </resources>
