+++ android/app/src/main/java/com/bvapps/easystocks/MainActivity.java
@@ -1,4 +1,4 @@
-package com.reactnativebase;
+package com.bvapps.easystocks;
 
 import com.facebook.react.ReactActivity;
 import com.facebook.react.ReactActivityDelegate;
@@ -9,7 +9,7 @@ public class MainActivity extends ReactActivity {
 
   @Override
   protected String getMainComponentName() {
-    return "ReactNativeBase";
+    return "EasyStocks";
   }
 
   @Override
