+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "ReactNativeBase",
-  "displayName": "ReactNativeBase"
+  "name": "EasyStocks",
+  "displayName": "Easy Stocks"
 }
