+++ ios/EasyStocks-Bridging-Header.h
