+++ ios/EasyStocks-tvOS/Info.plist
