+++ ios/EasyStocks-tvOSTests/Info.plist
