+++ ios/EasyStocks.xcodeproj/xcshareddata/xcschemes/ReactNativeBase-Develop.xcscheme
