+++ ios/EasyStocks.xcodeproj/xcshareddata/xcschemes/ReactNativeBase-Staging.xcscheme
