+++ ios/EasyStocks.xcworkspace/xcshareddata/IDEWorkspaceChecks.plist
