+++ ios/EasyStocks/AppDelegate.h
