+++ ios/EasyStocks/Images.xcassets/AppIcon.appiconset/Contents.json
