+++ ios/EasyStocks/Images.xcassets/Contents.json
