+++ ios/EasyStocks/main.m
