+++ ios/EasyStocksTests/EasyStocksTests.m
@@ -14,11 +14,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React Native!"
 
-@interface ReactNativeBaseTests : XCTestCase
+@interface EasyStocksTests : XCTestCase
 
 @end
 
-@implementation ReactNativeBaseTests
+@implementation EasyStocksTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL(^)(UIView *view))test
 {
