+++ ios/EasyStocksTests/Info.plist
